Added fix for PhoneNumber regex and update for applicants number (as field noOfApplicantCalculated) in the model schema
handler.ts: Added "$inc: { noOfApplicationCalculated : 1 }" as it updates the value of noOfApplicationCalculated when a new applicants are added. (it was not gettting incremented before)
validate.ts: Few validaton changes
a) in sectorsOther validation: remove the required for item validation inside the array as it is not a mandatory field in the front end and was causing errors while using the formcontrol method.
b) in phoneNumber validation: changed the phoneNumber regex as it was not allowing number like std (ex. 0614-123456), so changed the regex to include all digits from 0-9 in phone Number.
in enrollment1.ts: Added state and district field in the enrollment part as they are useful for the operation/analysis team.
index.ts: incresed the userLimit to higher value to ensure that user is able to perform a reasonable amound of actions, otherwise 429 error was given after using only few user clicks/actions.